chore: add v8 implementations
